Add layouts and functionality for location list fragment
Made a list item layout for the nearby locations recycler view and added
an adapter for view inflation. MainActivity now loads both map
and location list fragments on permission granted. Changed userRatingsTotal
from Int to Double to fix generated type.

+++ app/src/main/java/com/aoinc/group1_location_nearbyplaces/model/data/Result.kt
@@ -39,7 +39,7 @@ data class Result(
     @SerializedName("types")
     val types: List<String>,
     @SerializedName("user_ratings_total")
-    val userRatingsTotal: Int,
+    val userRatingsTotal: Double,
     @SerializedName("vicinity")
     val vicinity: String
 )
